refactor: restore `kdf.rs` and keep `derive_delta` there for now
