+++ chain-signatures/crypto/src/lib.rs
@@ -4,9 +4,9 @@ pub mod types;
 use k256::elliptic_curve::sec1::FromEncodedPoint;
 use k256::EncodedPoint;
 pub use kdf::{
-    check_ec_signature, derive_delta, derive_epsilon_canton, derive_epsilon_checkpoint,
-    derive_epsilon_eth, derive_epsilon_hydration, derive_epsilon_near, derive_epsilon_sol,
-    derive_key, into_signature, valid_signature, verify_signature, x_coordinate,
+    check_ec_signature, derive_epsilon_canton, derive_epsilon_checkpoint, derive_epsilon_eth,
+    derive_epsilon_hydration, derive_epsilon_near, derive_epsilon_sol, derive_key, into_signature,
+    valid_signature, verify_signature, x_coordinate,
 };
 pub use mpc_primitives::{PublicKey, ScalarExt};
